arch: migrate telegram_bot.graph.state to src.runtime.graph.state (#2049 slice 1)
First slice of the reverse-layering fix tracked under #1948 / #2045 /
#2049. Moves the LangGraph state schema (RAGState TypedDict) and the
make_initial_state factory out of telegram_bot/graph/state.py into the
runtime kernel src/runtime/graph/state.py so src/api and mini_app no
longer have to reach back into telegram_bot/ for them.

What this PR does:

- Adds src/runtime/graph/__init__.py and src/runtime/graph/state.py
  containing the canonical implementation copied verbatim from
  telegram_bot/graph/state.py (no behavior changes; same TypedDict
  fields, same make_initial_state defaults).
- Reduces telegram_bot/graph/state.py to a re-export shim
  from src.runtime.graph.state import RAGState, make_initial_state
  so every existing
  from telegram_bot.graph.state import … caller (28+ test files,
  telegram_bot/graph/graph.py, telegram_bot/graph/nodes/generate.py)
  keeps working without churn.
- Switches src/api/main.py:346 to import make_initial_state from
  src.runtime.graph.state, removing the reverse import from a hot path.
- Shrinks tests/data/known_layering_violations.json from 5 entries to 4
  for src/api/main.py: telegram_bot.graph.state is now compliant.

What this PR does NOT do (out of scope, follow-up work):

- The remaining 4 src/api/main.py imports (telegram_bot.graph.config,
  telegram_bot.graph.graph, telegram_bot.integrations.cache,
  telegram_bot.services.qdrant) still flow back into telegram_bot. They
  are tracked under #2045 (graph.config / graph.graph) and #2047
  (integrations.cache / services.qdrant) respectively. Issue #2049 stays
  open until the allowlist is empty and the test asserting
  empty allowlist is added (acceptance criterion).
- Mini App separability verification is not part of this slice.

Validation:

- uv run --frozen pytest tests/contract/test_layering_no_telegram_bot_imports_contract.py -q
  → 4 passed (allowlist drift detector picks up the shrunk JSON).
- uv run --frozen pytest tests/unit/graph/ tests/integration/test_graph_paths.py -q
  → 508 passed (back-compat shim works for every existing caller).
- ruff check + ruff format clean (auto-sorted the new import in
  src/api/main.py).

Refs #2049 #2045 #1948
