+++ tests/data/known_layering_violations.json
@@ -2,7 +2,6 @@
   "src/api/main.py": [
     "telegram_bot.graph.config",
     "telegram_bot.graph.graph",
-    "telegram_bot.graph.state",
     "telegram_bot.integrations.cache",
     "telegram_bot.services.qdrant"
   ]
